refacto(infrastructure): stop fastapi adapters from importing the composition root
api/infrastructure/fastapi/ was unimportable on its own: its __init__ pulled
AccessController, which imported api.dependencies — the module that also wires
every use case. Any import reaching the package before the composition root
failed with a circular import; the app only booted because api.app happened to
load them in the working order.

Low-level providers (session, redis client, key encoder/repository,
authenticated-user query) move to the leaf module api/infrastructure/
dependencies.py, so an adapter no longer depends on the root that assembles the
layer above it; api/dependencies.py keeps only use-case factories.

RequestContext and its usage recorder move back next to the FastAPI decorator
that drains them, and get_request_context/get_authenticated_user move down to
_requestcontext.py, next to the ContextVar they read. That removes the last edge
from api/infrastructure/dependencies.py to the fastapi package, which is what
lets __init__ re-export AccessController again.
