+++ api/infrastructure/fastapi/__init__.py
@@ -1,3 +1,12 @@
 from ._accesscontroler import AccessController
+from ._requestcontext import RequestContext, get_authenticated_user, get_request_context, request_context
+from ._requestcontextusagerecorder import RequestContextUsageRecorder
 
-__all__ = ["AccessController"]
+__all__ = [
+    "AccessController",
+    "RequestContext",
+    "RequestContextUsageRecorder",
+    "get_authenticated_user",
+    "get_request_context",
+    "request_context",
+]
