+++ api/infrastructure/fastapi/endpoints/admin/providers.py
@@ -8,7 +8,6 @@ from api.dependencies import (
     delete_provider_use_case_factory,
     get_one_provider_use_case_factory,
     get_providers_use_case_factory,
-    get_request_context,
     update_provider_use_case_factory,
 )
 from api.domain import SortOrder
@@ -22,8 +21,7 @@ from api.domain.provider.errors import (
     ProviderNotReachableError,
 )
 from api.domain.router.errors import RouterNotFoundError
-from api.infrastructure.context import RequestContext
-from api.infrastructure.fastapi import AccessController
+from api.infrastructure.fastapi import AccessController, RequestContext, get_request_context
 from api.infrastructure.fastapi.documentation import get_documentation_responses
 from api.infrastructure.fastapi.endpoints.admin import router
 from api.infrastructure.fastapi.endpoints.exceptions import (
