+++ api/infrastructure/fastapi/endpoints/admin/roles.py
@@ -6,7 +6,6 @@ from fastapi import Body, Depends, Path, Query, Security
 from api.dependencies import (
     create_role_use_case_factory,
     delete_role_use_case_factory,
-    get_request_context,
     get_role_use_case_factory,
     get_roles_use_case_factory,
     update_role_use_case_factory,
@@ -14,8 +13,7 @@ from api.dependencies import (
 from api.domain import SortField, SortOrder
 from api.domain.role.entities import Limit
 from api.domain.role.errors import RoleAlreadyExistsError, RoleHasUsersError, RoleNotFoundError
-from api.infrastructure.context import RequestContext
-from api.infrastructure.fastapi import AccessController
+from api.infrastructure.fastapi import AccessController, RequestContext, get_request_context
 from api.infrastructure.fastapi.documentation import get_documentation_responses
 from api.infrastructure.fastapi.endpoints.admin import router
 from api.infrastructure.fastapi.endpoints.exceptions import (
