+++ api/infrastructure/fastapi/endpoints/health.py
@@ -3,9 +3,8 @@ import logging
 from fastapi import APIRouter, Depends, Security
 from fastapi.responses import JSONResponse
 
-from api.dependencies import get_health_models_use_case_factory, get_request_context
-from api.infrastructure.context import RequestContext
-from api.infrastructure.fastapi import AccessController
+from api.dependencies import get_health_models_use_case_factory
+from api.infrastructure.fastapi import AccessController, RequestContext, get_request_context
 from api.infrastructure.fastapi.documentation import get_documentation_responses
 from api.infrastructure.fastapi.endpoints.exceptions import InternalServerHTTPException
 from api.infrastructure.fastapi.schemas.health import ModelHealthStatus, ModelsHealthResponse
