+++ api/tests/integ/test_me.py
@@ -39,7 +39,7 @@ def setup_model_and_user(client: TestClient):
         )
         role_id = create_role(router_id=router_id, client=client)
         user_id = create_user(role_id=role_id, client=client)
-        _, key = create_token(user_id=user_id, token_name=f"test-token-{dt.datetime.now().strftime("%Y%m%d%H%M%S")}", client=client)
+        _, key = create_token(user_id=user_id, token_name=f"test-token-{dt.datetime.now().strftime('%Y%m%d%H%M%S')}", client=client)
 
         yield key, process.model_name
     except Exception as e:
