+++ api/tests/integ/test_providers/test_mistral_text_image_to_text.py
@@ -38,7 +38,7 @@ def setup_mistral_image_text_to_text(client: TestClient):
         )
         role_id = create_role(router_id=router_id, client=client)
         user_id = create_user(role_id=role_id, client=client)
-        _, key = create_token(user_id=user_id, token_name=f"test-token-{dt.datetime.now().strftime("%Y%m%d%H%M%S")}", client=client)
+        _, key = create_token(user_id=user_id, token_name=f"test-token-{dt.datetime.now().strftime('%Y%m%d%H%M%S')}", client=client)
 
         yield key, process.model_name
     except Exception as e:
