+++ api/tests/integration/fastapi/test_accesscontroler.py
@@ -7,7 +7,7 @@ from jose import jwt
 import pytest
 
 from api.domain.role.entities import PermissionType
-from api.infrastructure.context import RequestContext, request_context
+from api.infrastructure.fastapi import RequestContext, request_context
 from api.infrastructure.fastapi._accesscontroler import AccessController
 from api.infrastructure.fastapi.endpoints.exceptions import (
     AccountExpiredHTTPException,
